+++ packages/at_cli_commons/example/bin/scan_example.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:args/args.dart';
 import 'package:at_cli_commons/at_cli_commons.dart';
 
 /// scan_example.dart
@@ -8,14 +9,21 @@ import 'package:at_cli_commons/at_cli_commons.dart';
 /// Create an atClient and list the identifiers of all the records stored
 /// in its atServer
 Future<void> main(List<String> args) async {
+  final ArgParser parser = CLIBase.createArgsParser(
+      namespace: 'example', addLegacyRootDomainArg: true);
   try {
-    var atClient = (await CLIBase.fromCommandLineArgs(args)).atClient;
-    var allDataKeys =
-        await atClient.getRemoteSecondary()!.executeCommand('scan\n');
+    var atClient =
+        (await CLIBase.fromCommandLineArgs(args, parser: parser)).atClient;
+    var allDataKeys = await atClient.getKeys(useRemoteAtServer: true);
     print(allDataKeys);
     exit(0);
-  } catch (e) {
+  } catch (e, st) {
     print(e);
-    print(CLIBase.argsParser.usage);
+    if (e is ArgumentError || e is FormatException) {
+      print(parser.usage);
+    } else {
+      print('Stack Trace:\n$st');
+    }
+    exit(1);
   }
 }