+++ packages/at_client/lib/src/rpc/at_rpc.dart
@@ -14,7 +14,6 @@ abstract class AtRpcCallbacks {
   Future<void> handleResponse(AtRpcResp response);
 }
 
-@experimental
 class AtRpcClient implements AtRpcCallbacks {
   static final AtSignLogger logger = AtSignLogger(' AtRpcClient ',
       loggingHandler: AtSignLogger.stdErrLoggingHandler);
@@ -40,6 +39,8 @@ class AtRpcClient implements AtRpcCallbacks {
       callbacks: this,
       allowList: {},
       allowAll: false,
+      isClient: true,
+      isServer: false,
     );
     rpc.start();
   }
@@ -103,7 +104,6 @@ class AtRpcClient implements AtRpcCallbacks {
 /// - Responder:
 /// ```
 /// ```
-@experimental
 class AtRpc {
   static final AtSignLogger logger = AtSignLogger('AtRpc');
 
@@ -167,6 +167,13 @@ class AtRpc {
   final bool isClient;
   final bool isServer;
 
+  /// Enables the mutex in [handleRequestNotification] to ensure redundancy
+  /// handling across all services that use [AtRpc].
+  ///
+  /// When enabled, it prevents multiple [AtRpc] responses from being sent for
+  /// the same request.
+  final bool enableRequestMutex;
+
   AtRpc({
     required this.atClient,
     required this.baseNameSpace,
@@ -177,6 +184,7 @@ class AtRpc {
     this.allowAll = false,
     this.isClient = true,
     this.isServer = true,
+    this.enableRequestMutex = false,
   }) {
     if (!isClient && !isServer) {
       throw IllegalArgumentException('isClient or isServer must be true');
@@ -281,13 +289,17 @@ class AtRpc {
   final Metadata _defaultMetaData = Metadata()
     ..isPublic = false
     ..isEncrypted = true
-    ..namespaceAware = true;
+    // namespaceAware IS SET TO FALSE FOR A REASON:
+    // https://github.com/atsign-foundation/at_client_sdk/pull/1670
+    ..namespaceAware = false;
 
   /// Not part of API, but visibleForTesting.
   /// Receives 'request' notifications, and
   /// - parses and validates
   /// - sends an [AtRpcRespType.nack] response if deserialization or validation fails
   /// - sends an [AtRpcRespType.nack] response otherwise
+  /// - Acquires session mutex if [enableRequestMutex] is true
+  /// - sends an [AtRpcRespType.nack] response if mutex cannot be acquired
   /// - calls [AtRpcCallbacks.handleRequest]
   /// - calls [sendResponse] with the response from [AtRpcCallbacks.handleRequest]
   @visibleForTesting
@@ -352,6 +364,12 @@ class AtRpc {
       return;
     }
 
+    bool mutexAcquired = enableRequestMutex &&
+        await _tryAcquireSessionMutex(requestId, notification.to);
+    if (enableRequestMutex && !mutexAcquired) {
+      return;
+    }
+
     // send ACK
     await sendResponse(notification, request, AtRpcResp.ack(request: request));
 
@@ -366,6 +384,34 @@ class AtRpc {
       logger.warning(st);
       await sendResponse(notification, request,
           AtRpcResp.nack(request: request, message: message));
+    }
+  }
+
+  Future<bool> _tryAcquireSessionMutex(int requestId, String atsign) async {
+    var mutexKey = AtKey.fromString('$requestId.session_mutexes.'
+        '$domainNameSpace.$rpcsNameSpace.$baseNameSpace$atsign')
+      ..metadata = (Metadata()
+        ..immutable = true // only one RPC client will succeed in doing this
+        ..ttl =
+            30000); // keeps the datastore clean + auto releases mutex after 30s
+    PutRequestOptions pro = PutRequestOptions()
+      ..shouldEncrypt = true
+      ..useRemoteAtServer = true;
+
+    try {
+      await atClient.put(mutexKey, 'lock', putRequestOptions: pro);
+      logger.shout(
+          '😎 Will handle request from $atsign; acquired mutex $mutexKey');
+      return true;
+    } catch (err) {
+      if (err.toString().toLowerCase().contains('immutable')) {
+        logger.shout('Will not handle request from $atsign'
+            '; could not acquire mutex $mutexKey');
+      } else {
+        logger.shout(
+            '🤷‍♂️ Will not handle; could not acquire mutex $mutexKey : $err');
+      }
+      return false;
     }
   }
 
@@ -461,11 +507,13 @@ class AtRpc {
           ..namespace = baseNameSpace
           ..metadata = _defaultMetaData;
 
+        var responseJson = jsonEncode(response.toJson());
+
         logger.info(
-            "Sending notification $responseAtKey with payload ${response.toJson()}");
+            "Sending notification $responseAtKey with payload $responseJson");
         await atClient.notificationService.notify(
             NotificationParams.forUpdate(responseAtKey,
-                value: jsonEncode(response.toJson()),
+                value: responseJson,
                 notificationExpiry: defaultNotificationExpiry),
             checkForFinalDeliveryStatus: false,
             waitForFinalDeliveryStatus: false);