+++ packages/at_utils/lib/at_utils.dart
@@ -1,5 +1,5 @@
-export 'package:at_utils/src/atsign_util.dart';
-export 'package:at_utils/src/atmetadata_utils.dart';
-export 'package:at_utils/src/config/app_config.dart';
 export 'package:at_utils/at_logger.dart';
+export 'package:at_utils/src/atmetadata_utils.dart';
+export 'package:at_utils/src/atsign_util.dart';
+export 'package:at_utils/src/config/app_config.dart';
 export 'package:at_utils/src/networking/pseudo_server_socket.dart';